don't tame RegExp, add simplistic tests for all three exports
'tap' and 'harden' cannot even load unless RegExp is left untamed, so disable
that taming for now.

This adds super-simple tests for Compartment, harden, and HandledPromise,
just enough to know that the basics work. I've disabled the Compartment test
when outside of SES because we don't have a suitable replacement for it yet.